Método listarUsuarios movido para UsuarioDAO
