+++ app/src/main/java/com/app/eventos/dao/UsuarioDAO.java
@@ -1,13 +1,20 @@
 package com.app.eventos.dao;
 
+import android.support.annotation.NonNull;
+
 import com.app.eventos.activities.LoginActivity;
 import com.app.eventos.dao.ConfiguracaoFirebase;
 import com.app.eventos.dao.ConfiguracaoFirebaseAuth;
 import com.app.eventos.model.Usuario;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.ValueEventListener;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class UsuarioDAO {
@@ -22,5 +29,26 @@ public class UsuarioDAO {
         Map<String, Object> childUpdates = new HashMap<>();
         childUpdates.put("/usuarios/" + idUser, usuarioValues);
         ConfiguracaoFirebase.getDatabaseReference().updateChildren(childUpdates);
+    }
+
+    public List<Usuario> listarUsuarios() {
+        final List<Usuario> usuarios = new ArrayList<>();
+
+        ConfiguracaoFirebase.getDatabaseReference().child("usuario").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                for (DataSnapshot objSnapshot : dataSnapshot.getChildren()) {
+                    Usuario usuario = objSnapshot.getValue(Usuario.class);
+                    usuarios.add(usuario);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+        });
+
+        return usuarios;
     }
 }
